refactor: 메서드 이름 변경 및 주석 삭제
