+++ ticket-service/src/main/java/com/flight_booking/ticket_service/infrastructure/messaging/TicketKafkaEndpoint.java
@@ -3,9 +3,8 @@ package com.flight_booking.ticket_service.infrastructure.messaging;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.flight_booking.common.application.dto.FlightCancelRequestDto;
 import com.flight_booking.common.application.dto.TicketRequestDto;
-import com.flight_booking.common.application.dto.TicketUpdateStatusRequestDto;
 import com.flight_booking.common.presentation.global.ApiResponse;
-import com.flight_booking.ticket_service.infrastructure.service.TicketService;
+import com.flight_booking.ticket_service.application.service.TicketService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.messaging.handler.annotation.Payload;
